refactor: move all handleEvent code to Vite context
