+++ packages/hydrogen/src/entry-server.tsx
@@ -3,28 +3,44 @@ import {
   Logger,
   logServerResponse,
   getLoggerFromContext,
-  log as noContextLogger,
 } from './utilities/log/log';
 import {getErrorMarkup} from './utilities/error';
 import {defer} from './utilities/defer';
-import type {ImportGlobEagerOutput, ServerHandler} from './types';
+import type {
+  RendererOptions,
+  StreamerOptions,
+  HydratorOptions,
+  ImportGlobEagerOutput,
+  ServerHandlerConfig,
+} from './types';
 import {Html} from './framework/Hydration/Html';
-import {Renderer, Hydrator, Streamer} from './types';
 import {ServerComponentResponse} from './framework/Hydration/ServerComponentResponse.server';
-import type {ServerComponentRequest} from './framework/Hydration/ServerComponentRequest.server';
+import {ServerComponentRequest} from './framework/Hydration/ServerComponentRequest.server';
 import {getCacheControlHeader} from './framework/cache';
 import {ServerRequestProvider} from './foundation/ServerRequestProvider';
 import type {ServerResponse} from 'http';
 import type {PassThrough as PassThroughType, Writable} from 'stream';
-import {getApiRouteFromURL, getApiRoutesFromPages} from './utilities/apiRoutes';
+import {
+  getApiRouteFromURL,
+  renderApiRoute,
+  getApiRoutesFromPages,
+} from './utilities/apiRoutes';
 import {ServerStateProvider} from './foundation/ServerStateProvider';
+import {isBotUA} from './utilities/bot-ua';
 import type {RealHelmetData} from './foundation/Helmet';
+
+import {setContext, setCache, RuntimeContext} from './framework/runtime';
+import {setConfig} from './framework/config';
+import type {IncomingMessage} from 'connect';
 
 import {
   ssrRenderToPipeableStream,
   ssrRenderToReadableStream,
   rscRenderToReadableStream,
   createFromReadableStream,
+  supportsReadableStream,
+  isStreamingSupported,
+  bufferReadableStream,
 } from './streaming.server';
 
 declare global {
@@ -43,439 +59,546 @@ const STREAM_ABORT_TIMEOUT_MS = 3000;
 
 const HTML_CONTENT_TYPE = 'text/html; charset=UTF-8';
 
-const renderHydrogen: ServerHandler = (App, {pages}) => {
-  /**
-   * The render function is responsible for turning the provided `App` into an HTML string,
-   * and returning any initial state that needs to be hydrated into the client version of the app.
-   * NOTE: This is currently only used for SEO bots or Worker runtime (where Stream is not yet supported).
-   */
-  const render: Renderer = async function (
-    url,
-    {request, template, nonce, dev}
+interface HandleEventOptions {
+  indexTemplate: string | ((url: string) => Promise<string>);
+  cache?: Cache;
+  streamableResponse?: ServerResponse;
+  dev?: boolean;
+  context?: RuntimeContext;
+  nonce?: string;
+  entrypoint: any;
+  event: any;
+  assetHandler?: (event: any, url: URL) => Promise<Response>;
+}
+
+export interface HandleRequest {
+  (request: Request | IncomingMessage, options: HandleEventOptions): Promise<
+    Response | undefined
+  >;
+}
+
+export const renderHydrogen = (App: any, {pages}: ServerHandlerConfig) => {
+  const handleRequest: HandleRequest = async function (
+    rawRequest,
+    {
+      indexTemplate,
+      streamableResponse,
+      dev,
+      cache,
+      context,
+      nonce,
+      assetHandler,
+      event,
+    }
   ) {
-    const {log, state, componentResponse} = setupCurrentRequest(url, request);
+    const request = new ServerComponentRequest(rawRequest);
+    const url = new URL(request.url);
+    const log = getLoggerFromContext(request);
+    const componentResponse = new ServerComponentResponse();
 
-    const {AppSSR, rscReadable} = buildAppSSR(
-      {
-        App,
-        state,
-        request,
-        response: componentResponse,
-        pages,
-        log,
-      },
-      {template}
-    );
+    if (/\.(png|jpe?g|gif|css|js|svg|ico|map)$/i.test(url.pathname)) {
+      if (assetHandler) return assetHandler(event, url);
+      log.warn('Asset file detected but assetHandler was not provided');
 
-    let [html, flight] = await Promise.all([
-      renderToBufferedString(AppSSR, {log, nonce}),
-      bufferReadableStream(rscReadable.getReader()),
-    ]);
-
-    const {headers, status, statusText} = getResponseOptions(componentResponse);
-
-    /**
-     * TODO: Also add `Vary` headers for `accept-language` and any other keys
-     * we want to shard our full-page cache for all Hydrogen storefronts.
-     */
-    headers[getCacheControlHeader({dev})] =
-      componentResponse.cacheControlHeader;
-
-    if (componentResponse.customBody) {
-      // This can be used to return sitemap.xml or any other custom response.
-
-      logServerResponse('ssr', log, request, status);
-
-      return new Response(await componentResponse.customBody, {
-        status,
-        statusText,
-        headers,
-      });
+      if (__WORKER__) return new Response(null, {status: 404});
+      if (streamableResponse) {
+        streamableResponse.statusCode = 404;
+        streamableResponse.end();
+        return;
+      }
     }
 
-    headers['Content-type'] = HTML_CONTENT_TYPE;
-    const {bodyAttributes, htmlAttributes, ...head} = extractHeadElements(
-      request.ctx.helmet
-    );
+    /**
+     * Inject the cache & context into the module loader so we can pull it out for subrequests.
+     */
+    setCache(cache);
+    setContext(context);
+    setConfig({dev});
 
-    html = html
-      .replace(
-        /<head>(.*?)<\/head>/s,
-        generateHeadTag(head as Record<string, any>)
-      )
-      .replace('<html', htmlAttributes ? `<html ${htmlAttributes}` : '$&')
-      .replace('<body', bodyAttributes ? `<body ${bodyAttributes}` : '$&')
-      .replace(
-        '</body>',
-        flight
-          ? `${flightContainer({init: true, nonce, chunk: flight})}</body>`
-          : '$&'
+    const isReactHydrationRequest = url.pathname === '/react';
+
+    const template =
+      typeof indexTemplate === 'function'
+        ? await indexTemplate(url.toString())
+        : indexTemplate;
+
+    // const {render, hydrate, stream, getApiRoute, log}: EntryServerHandler =
+    //   entrypoint.default || entrypoint;
+
+    // @ts-ignore
+    if (dev && !(render && hydrate && stream && getApiRoute)) {
+      throw new Error(
+        `entry-server.jsx could not be loaded. This likely occurred because of a Vite compilation error.\n` +
+          `Please check your server logs for more information.`
       );
+    }
+
+    if (!isReactHydrationRequest && pages) {
+      const apiRoute = getApiRoute(url, {pages});
+
+      // The API Route might have a default export, making it also a server component
+      // If it does, only render the API route if the request method is GET
+      if (
+        apiRoute &&
+        (!apiRoute.hasServerComponent || request.method !== 'GET')
+      ) {
+        return renderApiRoute(request, apiRoute, log);
+      }
+    }
+
+    const isStreamable =
+      !isBotUA(url, request.headers.get('user-agent')) &&
+      (!!streamableResponse || supportsReadableStream());
+
+    const params = {
+      App,
+      log,
+      dev,
+      pages,
+      nonce,
+      request,
+      template,
+      isStreamable,
+      componentResponse,
+      response: streamableResponse,
+    };
+
+    if (isReactHydrationRequest) {
+      return hydrate(url, params);
+    }
+
+    /**
+     * Stream back real-user responses, but for bots/etc,
+     * use `render` instead. This is because we need to inject <head>
+     * things for SEO reasons.
+     */
+    if (isStreamable) {
+      return stream(url, params);
+    }
+
+    return render(url, params);
+  };
+
+  return handleRequest;
+};
+
+function getApiRoute(url: URL, {pages}: {pages: ImportGlobEagerOutput}) {
+  const routes = getApiRoutesFromPages(pages);
+  return getApiRouteFromURL(url, routes);
+}
+
+/**
+ * The render function is responsible for turning the provided `App` into an HTML string,
+ * and returning any initial state that needs to be hydrated into the client version of the app.
+ * NOTE: This is currently only used for SEO bots or Worker runtime (where Stream is not yet supported).
+ */
+async function render(
+  url: URL,
+  {
+    App,
+    pages,
+    request,
+    componentResponse,
+    log,
+    template,
+    nonce,
+    dev,
+  }: RendererOptions
+) {
+  const state = {pathname: url.pathname, search: url.search};
+
+  const {AppSSR, rscReadable} = buildAppSSR(
+    {
+      App,
+      state,
+      request,
+      response: componentResponse,
+      pages,
+      log,
+    },
+    {template}
+  );
+
+  let [html, flight] = await Promise.all([
+    renderToBufferedString(AppSSR, {log, nonce}),
+    bufferReadableStream(rscReadable.getReader()),
+  ]);
+
+  const {headers, status, statusText} = getResponseOptions(componentResponse);
+
+  /**
+   * TODO: Also add `Vary` headers for `accept-language` and any other keys
+   * we want to shard our full-page cache for all Hydrogen storefronts.
+   */
+  headers[getCacheControlHeader({dev})] = componentResponse.cacheControlHeader;
+
+  if (componentResponse.customBody) {
+    // This can be used to return sitemap.xml or any other custom response.
 
     logServerResponse('ssr', log, request, status);
 
-    return new Response(html, {
+    return new Response(await componentResponse.customBody, {
       status,
       statusText,
       headers,
     });
-  };
+  }
 
-  /**
-   * Stream a response to the client. NOTE: This omits custom `<head>`
-   * information, so this method should not be used by crawlers.
-   */
-  const stream: Streamer = async function (
-    url: URL,
-    {request, response, template, nonce, dev}
-  ) {
-    const {log, state, componentResponse} = setupCurrentRequest(url, request);
-    log.trace('start stream');
+  headers['Content-type'] = HTML_CONTENT_TYPE;
+  const {bodyAttributes, htmlAttributes, ...head} = extractHeadElements(
+    request.ctx.helmet
+  );
 
-    const {AppSSR, rscReadable} = buildAppSSR(
-      {
-        App,
-        state,
-        request,
-        response: componentResponse,
-        log,
-        pages,
-      },
-      {
-        template,
-        htmlAttrs: {lang: 'en'},
-      }
+  html = html
+    .replace(
+      /<head>(.*?)<\/head>/s,
+      generateHeadTag(head as Record<string, any>)
+    )
+    .replace('<html', htmlAttributes ? `<html ${htmlAttributes}` : '$&')
+    .replace('<body', bodyAttributes ? `<body ${bodyAttributes}` : '$&')
+    .replace(
+      '</body>',
+      flight
+        ? `${flightContainer({init: true, nonce, chunk: flight})}</body>`
+        : '$&'
     );
 
-    const rscToScriptTagReadable = new ReadableStream({
-      start(controller) {
-        log.trace('rsc start chunks');
-        let init = true;
-        const encoder = new TextEncoder();
-        bufferReadableStream(rscReadable.getReader(), (chunk) => {
-          const scriptTag = flightContainer({init, chunk, nonce});
-          controller.enqueue(encoder.encode(scriptTag));
-          init = false;
-        }).then(() => {
-          log.trace('rsc finish chunks');
-          return controller.close();
-        });
-      },
-    });
+  logServerResponse('ssr', log, request, status);
 
-    let didError: Error | undefined;
+  return new Response(html, {
+    status,
+    statusText,
+    headers,
+  });
+}
 
-    if (__WORKER__) {
-      const deferredShouldReturnApp = defer<boolean>();
-      const encoder = new TextEncoder();
-      const transform = new TransformStream();
-      const writable = transform.writable.getWriter();
-      const responseOptions = {} as ResponseOptions;
+/**
+ * Stream a response to the client. NOTE: This omits custom `<head>`
+ * information, so this method should not be used by crawlers.
+ */
+async function stream(
+  url: URL,
+  {
+    App,
+    pages,
+    request,
+    response,
+    componentResponse,
+    log,
+    template,
+    nonce,
+    dev,
+  }: StreamerOptions
+) {
+  const state = {pathname: url.pathname, search: url.search};
+  log.trace('start stream');
 
-      const ssrReadable = ssrRenderToReadableStream(AppSSR, {
-        nonce,
-        onCompleteShell() {
-          log.trace('worker ready to stream');
-
-          Object.assign(
-            responseOptions,
-            getResponseOptions(componentResponse, didError)
-          );
-
-          /**
-           * TODO: This assumes `response.cache()` has been called _before_ any
-           * queries which might be caught behind Suspense. Clarify this or add
-           * additional checks downstream?
-           */
-          responseOptions.headers[getCacheControlHeader({dev})] =
-            componentResponse.cacheControlHeader;
-
-          if (isRedirect(responseOptions)) {
-            // Return redirects early without further rendering/streaming
-            return deferredShouldReturnApp.resolve(false);
-          }
-
-          if (!componentResponse.canStream()) return;
-
-          startWritingHtmlToStream(
-            responseOptions,
-            writable,
-            encoder,
-            dev ? didError : undefined
-          );
-
-          deferredShouldReturnApp.resolve(true);
-        },
-        async onCompleteAll() {
-          log.trace('worker complete stream');
-          if (componentResponse.canStream()) return;
-
-          Object.assign(
-            responseOptions,
-            getResponseOptions(componentResponse, didError)
-          );
-
-          if (isRedirect(responseOptions)) {
-            // Redirects found after any async code
-            return deferredShouldReturnApp.resolve(false);
-          }
-
-          if (componentResponse.customBody) {
-            writable.write(encoder.encode(await componentResponse.customBody));
-            return deferredShouldReturnApp.resolve(false);
-          }
-
-          startWritingHtmlToStream(
-            responseOptions,
-            writable,
-            encoder,
-            dev ? didError : undefined
-          );
-
-          deferredShouldReturnApp.resolve(true);
-        },
-        onError(error) {
-          didError = error;
-
-          if (dev && deferredShouldReturnApp.status === 'pending') {
-            writable.write(getErrorMarkup(error));
-          }
-
-          console.error(error);
-        },
-      });
-
-      if (await deferredShouldReturnApp.promise) {
-        let bufferedSsr = '';
-        let isPendingSsrWrite = false;
-        const writingSSR = bufferReadableStream(
-          ssrReadable.getReader(),
-          (chunk) => {
-            bufferedSsr += chunk;
-
-            if (!isPendingSsrWrite) {
-              isPendingSsrWrite = true;
-              setTimeout(() => {
-                isPendingSsrWrite = false;
-                // React can write fractional chunks synchronously.
-                // This timeout ensures we only write full HTML tags
-                // in order to allow RSC writing concurrently.
-                if (bufferedSsr) {
-                  writable.write(encoder.encode(bufferedSsr));
-                  bufferedSsr = '';
-                }
-              }, 0);
-            }
-          }
-        );
-
-        const writingRSC = bufferReadableStream(
-          rscToScriptTagReadable.getReader(),
-          (scriptTag) => writable.write(encoder.encode(scriptTag))
-        );
-
-        Promise.all([writingSSR, writingRSC]).then(() => {
-          // Last SSR write might be pending, delay closing the writable one tick
-          setTimeout(() => writable.close(), 0);
-          logServerResponse('str', log, request, responseOptions.status);
-        });
-      } else {
-        writable.close();
-        logServerResponse('str', log, request, responseOptions.status);
-      }
-
-      if (await isStreamingSupported()) {
-        return new Response(transform.readable, responseOptions);
-      }
-
-      const bufferedBody = await bufferReadableStream(
-        transform.readable.getReader()
-      );
-
-      return new Response(bufferedBody, responseOptions);
-    } else if (response) {
-      response.socket!.on('error', log.fatal);
-
-      const {pipe} = ssrRenderToPipeableStream(AppSSR, {
-        nonce,
-        onCompleteShell() {
-          log.trace('node ready to stream');
-          /**
-           * TODO: This assumes `response.cache()` has been called _before_ any
-           * queries which might be caught behind Suspense. Clarify this or add
-           * additional checks downstream?
-           */
-          response.setHeader(
-            getCacheControlHeader({dev}),
-            componentResponse.cacheControlHeader
-          );
-
-          writeHeadToServerResponse(response, componentResponse, log, didError);
-
-          logServerResponse('str', log, request, response.statusCode);
-
-          if (isRedirect(response)) {
-            // Return redirects early without further rendering/streaming
-            return response.end();
-          }
-
-          if (!componentResponse.canStream()) return;
-
-          startWritingHtmlToServerResponse(
-            response,
-            dev ? didError : undefined
-          );
-
-          setTimeout(() => {
-            log.trace('node pipe response');
-            pipe(response);
-          }, 0);
-
-          bufferReadableStream(rscToScriptTagReadable.getReader(), (chunk) => {
-            log.trace('rsc chunk');
-            return response.write(chunk);
-          });
-        },
-        async onCompleteAll() {
-          log.trace('node complete stream');
-          clearTimeout(streamTimeout);
-
-          if (componentResponse.canStream() || response.writableEnded) return;
-
-          writeHeadToServerResponse(response, componentResponse, log, didError);
-
-          logServerResponse('str', log, request, response.statusCode);
-
-          if (isRedirect(response)) {
-            // Redirects found after any async code
-            return response.end();
-          }
-
-          if (componentResponse.customBody) {
-            return response.end(await componentResponse.customBody);
-          }
-
-          startWritingHtmlToServerResponse(
-            response,
-            dev ? didError : undefined
-          );
-
-          bufferReadableStream(rscToScriptTagReadable.getReader()).then(
-            (scriptTags) => {
-              // Piping ends the response so script tags
-              // must be written before that.
-              response.write(scriptTags);
-              pipe(response);
-            }
-          );
-        },
-        onError(error: any) {
-          didError = error;
-
-          if (dev && response.headersSent) {
-            // Calling write would flush headers automatically.
-            // Delay this error until headers are properly sent.
-            response.write(getErrorMarkup(error));
-          }
-
-          log.error(error);
-        },
-      });
-
-      const streamTimeout = setTimeout(() => {
-        log.warn(
-          `The app failed to stream after ${STREAM_ABORT_TIMEOUT_MS} ms`
-        );
-      }, STREAM_ABORT_TIMEOUT_MS);
-    }
-  };
-
-  /**
-   * Stream a hydration response to the client.
-   */
-  const hydrate: Hydrator = async function (
-    url: URL,
-    {request, response, isStreamable}
-  ) {
-    const {log, state, componentResponse} = setupCurrentRequest(url, request);
-
-    const {AppRSC} = buildAppRSC({
+  const {AppSSR, rscReadable} = buildAppSSR(
+    {
       App,
       state,
       request,
       response: componentResponse,
       log,
       pages,
+    },
+    {
+      template,
+      htmlAttrs: {lang: 'en'},
+    }
+  );
+
+  const rscToScriptTagReadable = new ReadableStream({
+    start(controller) {
+      log.trace('rsc start chunks');
+      let init = true;
+      const encoder = new TextEncoder();
+      bufferReadableStream(rscReadable.getReader(), (chunk) => {
+        const scriptTag = flightContainer({init, chunk, nonce});
+        controller.enqueue(encoder.encode(scriptTag));
+        init = false;
+      }).then(() => {
+        log.trace('rsc finish chunks');
+        return controller.close();
+      });
+    },
+  });
+
+  let didError: Error | undefined;
+
+  if (__WORKER__) {
+    const deferredShouldReturnApp = defer<boolean>();
+    const encoder = new TextEncoder();
+    const transform = new TransformStream();
+    const writable = transform.writable.getWriter();
+    const responseOptions = {} as ResponseOptions;
+
+    const ssrReadable = ssrRenderToReadableStream(AppSSR, {
+      nonce,
+      onCompleteShell() {
+        log.trace('worker ready to stream');
+
+        Object.assign(
+          responseOptions,
+          getResponseOptions(componentResponse, didError)
+        );
+
+        /**
+         * TODO: This assumes `response.cache()` has been called _before_ any
+         * queries which might be caught behind Suspense. Clarify this or add
+         * additional checks downstream?
+         */
+        responseOptions.headers[getCacheControlHeader({dev})] =
+          componentResponse.cacheControlHeader;
+
+        if (isRedirect(responseOptions)) {
+          // Return redirects early without further rendering/streaming
+          return deferredShouldReturnApp.resolve(false);
+        }
+
+        if (!componentResponse.canStream()) return;
+
+        startWritingHtmlToStream(
+          responseOptions,
+          writable,
+          encoder,
+          dev ? didError : undefined
+        );
+
+        deferredShouldReturnApp.resolve(true);
+      },
+      async onCompleteAll() {
+        log.trace('worker complete stream');
+        if (componentResponse.canStream()) return;
+
+        Object.assign(
+          responseOptions,
+          getResponseOptions(componentResponse, didError)
+        );
+
+        if (isRedirect(responseOptions)) {
+          // Redirects found after any async code
+          return deferredShouldReturnApp.resolve(false);
+        }
+
+        if (componentResponse.customBody) {
+          writable.write(encoder.encode(await componentResponse.customBody));
+          return deferredShouldReturnApp.resolve(false);
+        }
+
+        startWritingHtmlToStream(
+          responseOptions,
+          writable,
+          encoder,
+          dev ? didError : undefined
+        );
+
+        deferredShouldReturnApp.resolve(true);
+      },
+      onError(error) {
+        didError = error;
+
+        if (dev && deferredShouldReturnApp.status === 'pending') {
+          writable.write(getErrorMarkup(error));
+        }
+
+        console.error(error);
+      },
     });
 
-    if (__WORKER__) {
-      const rscReadable = rscRenderToReadableStream(AppRSC);
+    if (await deferredShouldReturnApp.promise) {
+      let bufferedSsr = '';
+      let isPendingSsrWrite = false;
+      const writingSSR = bufferReadableStream(
+        ssrReadable.getReader(),
+        (chunk) => {
+          bufferedSsr += chunk;
 
-      if (isStreamable && (await isStreamingSupported())) {
-        logServerResponse('rsc', log, request, 200);
-        return new Response(rscReadable);
-      }
-
-      // Note: CFW does not support reader.piteTo nor iterable syntax
-      const bufferedBody = await bufferReadableStream(rscReadable.getReader());
-
-      logServerResponse('rsc', log, request, 200);
-
-      return new Response(bufferedBody);
-    } else if (response) {
-      response.socket!.on('error', log.fatal);
-
-      const rscWriter = await import(
-        // @ts-ignore
-        '@shopify/hydrogen/vendor/react-server-dom-vite/writer.node.server'
+          if (!isPendingSsrWrite) {
+            isPendingSsrWrite = true;
+            setTimeout(() => {
+              isPendingSsrWrite = false;
+              // React can write fractional chunks synchronously.
+              // This timeout ensures we only write full HTML tags
+              // in order to allow RSC writing concurrently.
+              if (bufferedSsr) {
+                writable.write(encoder.encode(bufferedSsr));
+                bufferedSsr = '';
+              }
+            }, 0);
+          }
+        }
       );
 
-      const stream = rscWriter
-        .renderToPipeableStream(AppRSC)
-        .pipe(response) as Writable;
+      const writingRSC = bufferReadableStream(
+        rscToScriptTagReadable.getReader(),
+        (scriptTag) => writable.write(encoder.encode(scriptTag))
+      );
 
-      stream.on('finish', function () {
-        logServerResponse('rsc', log, request, response!.statusCode);
+      Promise.all([writingSSR, writingRSC]).then(() => {
+        // Last SSR write might be pending, delay closing the writable one tick
+        setTimeout(() => writable.close(), 0);
+        logServerResponse('str', log, request, responseOptions.status);
       });
+    } else {
+      writable.close();
+      logServerResponse('str', log, request, responseOptions.status);
     }
-  };
 
-  function getApiRoute(url: URL) {
-    const routes = getApiRoutesFromPages(pages);
-    return getApiRouteFromURL(url, routes);
+    if (await isStreamingSupported()) {
+      return new Response(transform.readable, responseOptions);
+    }
+
+    const bufferedBody = await bufferReadableStream(
+      transform.readable.getReader()
+    );
+
+    return new Response(bufferedBody, responseOptions);
+  } else if (response) {
+    response.socket!.on('error', log.fatal);
+
+    const {pipe} = ssrRenderToPipeableStream(AppSSR, {
+      nonce,
+      onCompleteShell() {
+        log.trace('node ready to stream');
+        /**
+         * TODO: This assumes `response.cache()` has been called _before_ any
+         * queries which might be caught behind Suspense. Clarify this or add
+         * additional checks downstream?
+         */
+        response.setHeader(
+          getCacheControlHeader({dev}),
+          componentResponse.cacheControlHeader
+        );
+
+        writeHeadToServerResponse(response, componentResponse, log, didError);
+
+        logServerResponse('str', log, request, response.statusCode);
+
+        if (isRedirect(response)) {
+          // Return redirects early without further rendering/streaming
+          return response.end();
+        }
+
+        if (!componentResponse.canStream()) return;
+
+        startWritingHtmlToServerResponse(response, dev ? didError : undefined);
+
+        setTimeout(() => {
+          log.trace('node pipe response');
+          pipe(response);
+        }, 0);
+
+        bufferReadableStream(rscToScriptTagReadable.getReader(), (chunk) => {
+          log.trace('rsc chunk');
+          return response.write(chunk);
+        });
+      },
+      async onCompleteAll() {
+        log.trace('node complete stream');
+        clearTimeout(streamTimeout);
+
+        if (componentResponse.canStream() || response.writableEnded) return;
+
+        writeHeadToServerResponse(response, componentResponse, log, didError);
+
+        logServerResponse('str', log, request, response.statusCode);
+
+        if (isRedirect(response)) {
+          // Redirects found after any async code
+          return response.end();
+        }
+
+        if (componentResponse.customBody) {
+          return response.end(await componentResponse.customBody);
+        }
+
+        startWritingHtmlToServerResponse(response, dev ? didError : undefined);
+
+        bufferReadableStream(rscToScriptTagReadable.getReader()).then(
+          (scriptTags) => {
+            // Piping ends the response so script tags
+            // must be written before that.
+            response.write(scriptTags);
+            pipe(response);
+          }
+        );
+      },
+      onError(error: any) {
+        didError = error;
+
+        if (dev && response.headersSent) {
+          // Calling write would flush headers automatically.
+          // Delay this error until headers are properly sent.
+          response.write(getErrorMarkup(error));
+        }
+
+        log.error(error);
+      },
+    });
+
+    const streamTimeout = setTimeout(() => {
+      log.warn(`The app failed to stream after ${STREAM_ABORT_TIMEOUT_MS} ms`);
+    }, STREAM_ABORT_TIMEOUT_MS);
   }
+}
 
-  return {
-    render,
-    stream,
-    hydrate,
-    getApiRoute,
-    log: noContextLogger,
-  };
-};
-
-async function bufferReadableStream(
-  reader: ReadableStreamDefaultReader,
-  cb?: (chunk: string) => void
+/**
+ * Stream a hydration response to the client.
+ */
+async function hydrate(
+  url: URL,
+  {
+    App,
+    pages,
+    request,
+    response,
+    componentResponse,
+    isStreamable,
+    log,
+  }: HydratorOptions
 ) {
-  const decoder = new TextDecoder();
-  let result = '';
+  const state = JSON.parse(url.searchParams.get('state') || '{}');
 
-  while (true) {
-    const {done, value} = await reader.read();
-    if (done) break;
+  const {AppRSC} = buildAppRSC({
+    App,
+    state,
+    request,
+    response: componentResponse,
+    log,
+    pages,
+  });
 
-    const stringValue =
-      typeof value === 'string' ? value : decoder.decode(value);
+  if (__WORKER__) {
+    const rscReadable = rscRenderToReadableStream(AppRSC);
 
-    result += stringValue;
-
-    if (cb) {
-      cb(stringValue);
+    if (isStreamable && (await isStreamingSupported())) {
+      logServerResponse('rsc', log, request, 200);
+      return new Response(rscReadable);
     }
-  }
 
-  return result;
+    // Note: CFW does not support reader.piteTo nor iterable syntax
+    const bufferedBody = await bufferReadableStream(rscReadable.getReader());
+
+    logServerResponse('rsc', log, request, 200);
+
+    return new Response(bufferedBody);
+  } else if (response) {
+    response.socket!.on('error', log.fatal);
+
+    const rscWriter = await import(
+      // @ts-ignore
+      '@shopify/hydrogen/vendor/react-server-dom-vite/writer.node.server'
+    );
+
+    const stream = rscWriter
+      .renderToPipeableStream(AppRSC)
+      .pipe(response) as Writable;
+
+    stream.on('finish', function () {
+      logServerResponse('rsc', log, request, response!.statusCode);
+    });
+  }
 }
 
 type BuildAppOptions = {
@@ -764,39 +887,4 @@ function flightContainer({
   }
 
   return script + '</script>';
-}
-
-let cachedStreamingSupport: boolean;
-async function isStreamingSupported() {
-  if (cachedStreamingSupport === undefined) {
-    try {
-      const rs = new ReadableStream({
-        start(controller) {
-          controller.close();
-        },
-      });
-
-      // This will throw in CFW until streaming
-      // is supported. It works in Miniflare.
-      await new Response(rs).text();
-
-      cachedStreamingSupport = true;
-    } catch (_) {
-      cachedStreamingSupport = false;
-    }
-  }
-
-  return cachedStreamingSupport;
-}
-
-function setupCurrentRequest(url: URL, request: ServerComponentRequest) {
-  const log = getLoggerFromContext(request);
-  const state =
-    url.pathname === '/react'
-      ? JSON.parse(url.searchParams.get('state') || '{}')
-      : {pathname: url.pathname, search: url.search};
-
-  const componentResponse = new ServerComponentResponse();
-
-  return {log, state, componentResponse};
 }
