+++ packages/hydrogen/src/framework/middleware.ts
@@ -106,6 +106,7 @@ export function hydrogenMiddleware({
           streamableResponse: response,
           dev,
           cache,
+          event: {},
         }
       );
