+++ packages/hydrogen/src/streaming.server.ts
@@ -29,3 +29,59 @@ export const ssrRenderToReadableStream = _ssrRenderToReadableStream as (
     onError?: (error: Error) => void;
   }
 ) => ReadableStream<Uint8Array>;
+
+export function supportsReadableStream() {
+  try {
+    new ReadableStream();
+    return true;
+  } catch (_) {
+    return false;
+  }
+}
+
+let cachedStreamingSupport: boolean;
+export async function isStreamingSupported() {
+  if (cachedStreamingSupport === undefined) {
+    try {
+      const rs = new ReadableStream({
+        start(controller) {
+          controller.close();
+        },
+      });
+
+      // This will throw in CFW until streaming
+      // is supported. It works in Miniflare.
+      await new Response(rs).text();
+
+      cachedStreamingSupport = true;
+    } catch (_) {
+      cachedStreamingSupport = false;
+    }
+  }
+
+  return cachedStreamingSupport;
+}
+
+export async function bufferReadableStream(
+  reader: ReadableStreamDefaultReader,
+  cb?: (chunk: string) => void
+) {
+  const decoder = new TextDecoder();
+  let result = '';
+
+  while (true) {
+    const {done, value} = await reader.read();
+    if (done) break;
+
+    const stringValue =
+      typeof value === 'string' ? value : decoder.decode(value);
+
+    result += stringValue;
+
+    if (cb) {
+      cb(stringValue);
+    }
+  }
+
+  return result;
+}
