+++ packages/hydrogen/src/types.ts
@@ -1,46 +1,32 @@
 import {ServerResponse} from 'http';
+import type {Logger} from './utilities/log/log';
 import type {ServerComponentRequest} from './framework/Hydration/ServerComponentRequest.server';
+import type {ServerComponentResponse} from './framework/Hydration/ServerComponentResponse.server';
 import type {Metafield, Image, MediaContentType} from './graphql/types/types';
-import {ApiRouteMatch} from './utilities/apiRoutes';
-import {Logger} from './utilities/log/log';
 
-export type Renderer = (
-  url: URL,
-  options: {
-    request: ServerComponentRequest;
-    template: string;
-    nonce?: string;
-    dev?: boolean;
-  }
-) => Promise<Response>;
-
-export type Streamer = (
-  url: URL,
-  options: {
-    request: ServerComponentRequest;
-    response?: ServerResponse;
-    template: string;
-    nonce?: string;
-    dev?: boolean;
-  }
-) => void;
-
-export type Hydrator = (
-  url: URL,
-  options: {
-    request: ServerComponentRequest;
-    response?: ServerResponse;
-    isStreamable: boolean;
-    dev?: boolean;
-  }
-) => void;
-
-export type EntryServerHandler = {
-  render: Renderer;
-  stream: Streamer;
-  hydrate: Hydrator;
-  getApiRoute: (url: URL) => ApiRouteMatch | null;
+type CommonOptions = {
+  App: any;
+  pages?: ImportGlobEagerOutput;
+  request: ServerComponentRequest;
+  componentResponse: ServerComponentResponse;
   log: Logger;
+  dev?: boolean;
+};
+
+export type RendererOptions = CommonOptions & {
+  template: string;
+  nonce?: string;
+};
+
+export type StreamerOptions = CommonOptions & {
+  response?: ServerResponse;
+  template: string;
+  nonce?: string;
+};
+
+export type HydratorOptions = CommonOptions & {
+  response?: ServerResponse;
+  isStreamable: boolean;
 };
 
 export type ShopifyConfig = {
@@ -67,11 +53,6 @@ export type ServerHandlerConfig = {
 export type ClientHandlerConfig = {
   shopifyConfig: ShopifyConfig;
 };
-
-export type ServerHandler = (
-  App: any,
-  config: ServerHandlerConfig
-) => EntryServerHandler;
 
 export type ClientHandler = (
   App: any,
